% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postBetaVar.r
\name{postBetaVar}
\alias{postBetaVar}
\title{Posterior Beta Variance}
\usage{
postBetaVar(priorVar, ss.x, trueVar)
}
\arguments{
\item{priorVar}{Variance of prior distribution}

\item{ss.x}{Sum of square deviations of X from its mean}

\item{trueVar}{Hypothesised true variance}
}
\description{
This function calculates posterior beta coefficeint variance prior var, SSX and true variance
}
