% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postBetaVar.r
\name{postBetaVar}
\alias{postBetaVar}
\title{Posterior Beta Variance}
\usage{
postBetaVar(priorVar, ss.x, trueVar)
}
\arguments{
\item{priorVar}{Variance of prior distribution}

\item{ss.x}{Sum of square deviations of X from its mean}

\item{trueVar}{Hypothesised true variance}
}
\description{
This function calculates posterior beta coefficeint variance prior var, SSX and true variance
}
\examples{
HR.x <- c(94,96,94,95,104,106,108,113,115,121,131)
O2.y <- c(0.47,0.75,0.83,0.98,1.18,1.29,1.40,1.60,1.75,1.90,2.23)
SS.x1 <- t(as.vector(HR.x-mean(HR.x)))\%*\%(as.vector(HR.x-mean(HR.x))); SS.x1
PostVar <- postBetaVar(1,SS.x1,.0169); PostVar
}
